// types/restcountries.ts
export interface RestCountry {
  name: Name;
  tld?: string[];
  cca2: string;
  ccn3?: string;
  cca3: string;
  independent?: boolean;
  status: string;
  unMember: boolean;
  currencies?: Currencies;
  idd: Idd;
  capital?: string[];
  altSpellings: string[];
  region: string;
  languages?: Languages;
  translations: Translations;
  latlng: number[];
  landlocked: boolean;
  area: number;
  demonyms?: Demonyms;
  flag: string;
  maps: Maps;
  population: number;
  car: Car;
  timezones: string[];
  continents: string[];
  flags: Flags;
  coatOfArms: CoatOfArms;
  startOfWeek: string;
  capitalInfo: CapitalInfo;
  cioc?: string;
  subregion?: string;
  fifa?: string;
  borders?: string[];
  gini?: Gini;
  postalCode?: PostalCode;
}

export interface Name {
  common: string;
  official: string;
  nativeName?: NativeName;
}

export interface NativeName {
  eng?: Eng;
  spa?: Spa;
  mri?: Mri;
  nzs?: Nzs;
  ara?: Ara;
  heb?: Heb;
  deu?: Deu;
  fra?: Fra;
  nld?: Nld;
  kon?: Kon;
  lin?: Lin;
  smo?: Smo;
  por?: Por;
  mon?: Mon;
  bwg?: Bwg;
  kck?: Kck;
  khi?: Khi;
  ndc?: Ndc;
  nde?: Nde;
  nya?: Nya;
  sna?: Sna;
  sot?: Sot;
  toi?: Toi;
  tsn?: Tsn;
  tso?: Tso;
  ven?: Ven;
  xho?: Xho;
  zib?: Zib;
  gil?: Gil;
  ton?: Ton;
  cha?: Cha;
  zdj?: Zdj;
  msa?: Msa;
  pol?: Pol;
  jam?: Jam;
  dzo?: Dzo;
  mya?: Mya;
  bjz?: Bjz;
  srp?: Srp;
  sqi?: Sqi;
  fas?: Fas;
  div?: Div;
  mlt?: Mlt;
  rus?: Rus;
  tgk?: Tgk;
  crs?: Crs;
  aym?: Aym;
  grn?: Grn;
  que?: Que;
  jpn?: Jpn;
  nrf?: Nrf;
  hye?: Hye;
  ron?: Ron;
  kaz?: Kaz;
  slk?: Slk;
  lit?: Lit;
  kor?: Kor;
  swa?: Swa;
  tir?: Tir;
  sag?: Sag;
  swe?: Swe;
  nno?: Nno;
  nob?: Nob;
  smi?: Smi;
  pap?: Pap;
  gle?: Gle;
  tuk?: Tuk;
  bar?: Bar;
  zho?: Zho;
  ita?: Ita;
  lat?: Lat;
  vie?: Vie;
  tet?: Tet;
  est?: Est;
  kin?: Kin;
  niu?: Niu;
  bis?: Bis;
  ces?: Ces;
  som?: Som;
  lua?: Lua;
  dan?: Dan;
  fil?: Fil;
  pau?: Pau;
  lav?: Lav;
  bel?: Bel;
  tvl?: Tvl;
  hat?: Hat;
  tha?: Tha;
  ber?: Ber;
  mey?: Mey;
  nau?: Nau;
  ltz?: Ltz;
  isl?: Isl;
  mah?: Mah;
  slv?: Slv;
  amh?: Amh;
  bos?: Bos;
  hrv?: Hrv;
  ell?: Ell;
  tur?: Tur;
  kir?: Kir;
  arc?: Arc;
  ckb?: Ckb;
  khm?: Khm;
  kal?: Kal;
  bul?: Bul;
  afr?: Afr;
  her?: Her;
  hgm?: Hgm;
  kwn?: Kwn;
  loz?: Loz;
  ndo?: Ndo;
  glv?: Glv;
  ukr?: Ukr;
  ssw?: Ssw;
  cnr?: Cnr;
  hin?: Hin;
  tam?: Tam;
  mlg?: Mlg;
  nor?: Nor;
  pih?: Pih;
  hmo?: Hmo;
  tpi?: Tpi;
  urd?: Urd;
  kat?: Kat;
  nep?: Nep;
  fij?: Fij;
  hif?: Hif;
  gsw?: Gsw;
  roh?: Roh;
  hun?: Hun;
  ind?: Ind;
  lao?: Lao;
  cat?: Cat;
  run?: Run;
  nbl?: Nbl;
  nso?: Nso;
  zul?: Zul;
  cal?: Cal;
  mkd?: Mkd;
  nfr?: Nfr;
  fao?: Fao;
  uzb?: Uzb;
  sin?: Sin;
  ben?: Ben;
  prs?: Prs;
  pus?: Pus;
  rar?: Rar;
  fin?: Fin;
  tkl?: Tkl;
  pov?: Pov;
  aze?: Aze;
  mfe?: Mfe;
}
// ...
export interface Currencies {
  XCD?: Xcd;
  NIO?: Nio;
  NZD?: Nzd;
  ILS?: Ils;
  EUR?: Eur;
  XAF?: Xaf;
  USD?: Usd;
  MNT?: Mnt;
  ZWL?: Zwl;
  GHS?: Ghs;
  AUD?: Aud;
  KID?: Kid;
  TOP?: Top;
  KMF?: Kmf;
  BND?: Bnd;
  SGD?: Sgd;
  KYD?: Kyd;
  PLN?: Pln;
  JMD?: Jmd;
  LRD?: Lrd;
  VES?: Ves;
  BTN?: Btn;
  INR?: Inr;
  MMK?: Mmk;
  BZD?: Bzd;
  RSD?: Rsd;
  BRL?: Brl;
  ALL?: All;
  IRR?: Irr;
  SSP?: Ssp;
  MVR?: Mvr;
  KWD?: Kwd;
  BSD?: Bsd;
  NGN?: Ngn;
  GMD?: Gmd;
  TJS?: Tjs;
  GTQ?: Gtq;
  SCR?: Scr;
  BMD?: Bmd;
  CLP?: Clp;
  BOB?: Bob;
  JPY?: Jpy;
  GBP?: Gbp;
  JEP?: Jep;
  WST?: Wst;
  AMD?: Amd;
  FKP?: Fkp;
  XOF?: Xof;
  MDL?: Mdl;
  KZT?: Kzt;
  AOA?: Aoa;
  MRU?: Mru;
  JOD?: Jod;
  KRW?: Krw;
  TZS?: Tzs;
  ERN?: Ern;
  NOK?: Nok;
  ANG?: Ang;
  PAB?: Pab;
  MZN?: Mzn;
  TMT?: Tmt;
  HKD?: Hkd;
  UYU?: Uyu;
  VND?: Vnd;
  GYD?: Gyd;
  TTD?: Ttd;
  RON?: Ron2;
  RWF?: Rwf;
  KES?: Kes;
  SHP?: Shp;
  VUV?: Vuv;
  XPF?: Xpf;
  CZK?: Czk;
  SOS?: Sos;
  CDF?: Cdf;
  SRD?: Srd;
  BHD?: Bhd;
  GNF?: Gnf;
  DKK?: Dkk;
  GIP?: Gip;
  PHP?: Php;
  BYN?: Byn;
  HNL?: Hnl;
  TVD?: Tvd;
  HTG?: Htg;
  THB?: Thb;
  CRC?: Crc;
  DZD?: Dzd;
  MAD?: Mad;
  ARS?: Ars;
  ISK?: Isk;
  LSL?: Lsl;
  ZAR?: Zar;
  MOP?: Mop;
  OMR?: Omr;
  ETB?: Etb;
  BAM?: Bam;
  SAR?: Sar;
  MWK?: Mwk;
  CAD?: Cad;
  STN?: Stn;
  MYR?: Myr;
  RUB?: Rub;
  KGS?: Kgs;
  CUC?: Cuc;
  CUP?: Cup;
  SEK?: Sek;
  IQD?: Iqd;
  KHR?: Khr;
  BGN?: Bgn;
  AED?: Aed;
  NAD?: Nad;
  IMP?: Imp;
  UAH?: Uah;
  PYG?: Pyg;
  SZL?: Szl;
  SYP?: Syp;
  MGA?: Mga;
  QAR?: Qar;
  DOP?: Dop;
  PGK?: Pgk;
  LBP?: Lbp;
  BWP?: Bwp;
  PKR?: Pkr;
  GEL?: Gel;
  NPR?: Npr;
  CHF?: Chf;
  FJD?: Fjd;
  SDG?: Sdg;
  SLL?: Sll;
  HUF?: Huf;
  TWD?: Twd;
  BBD?: Bbd;
  TND?: Tnd;
  IDR?: Idr;
  CVE?: Cve;
  LAK?: Lak;
  UGX?: Ugx;
  BIF?: Bif;
  LYD?: Lyd;
  MXN?: Mxn;
  MKD?: Mkd2;
  CNY?: Cny;
  YER?: Yer;
  GGP?: Ggp;
  SBD?: Sbd;
  FOK?: Fok;
  UZS?: Uzs;
  EGP?: Egp;
  LKR?: Lkr;
  BDT?: Bdt;
  PEN?: Pen;
  TRY?: Try;
  AFN?: Afn;
  AWG?: Awg;
  CKD?: Ckd;
  ZMW?: Zmw;
  AZN?: Azn;
  DJF?: Djf;
  KPW?: Kpw;
  MUR?: Mur;
  COP?: Cop;
}

export interface Xcd {
  name: string;
  symbol: string;
}

export interface Nio {
  name: string;
  symbol: string;
}

export interface Nzd {
  name: string;
  symbol: string;
}

export interface Ils {
  name: string;
  symbol: string;
}

export interface Eur {
  name: string;
  symbol: string;
}

export interface Xaf {
  name: string;
  symbol: string;
}

export interface Usd {
  name: string;
  symbol: string;
}

export interface Mnt {
  name: string;
  symbol: string;
}

export interface Zwl {
  name: string;
  symbol: string;
}

export interface Ghs {
  name: string;
  symbol: string;
}

export interface Aud {
  name: string;
  symbol: string;
}

export interface Kid {
  name: string;
  symbol: string;
}

export interface Top {
  name: string;
  symbol: string;
}

export interface Kmf {
  name: string;
  symbol: string;
}

export interface Bnd {
  name: string;
  symbol: string;
}

export interface Sgd {
  name: string;
  symbol: string;
}

export interface Kyd {
  name: string;
  symbol: string;
}

export interface Pln {
  name: string;
  symbol: string;
}

export interface Jmd {
  name: string;
  symbol: string;
}

export interface Lrd {
  name: string;
  symbol: string;
}

export interface Ves {
  name: string;
  symbol: string;
}

export interface Btn {
  name: string;
  symbol: string;
}

export interface Inr {
  name: string;
  symbol: string;
}

export interface Mmk {
  name: string;
  symbol: string;
}

export interface Bzd {
  name: string;
  symbol: string;
}

export interface Rsd {
  name: string;
  symbol: string;
}

export interface Brl {
  name: string;
  symbol: string;
}

export interface All {
  name: string;
  symbol: string;
}

export interface Irr {
  name: string;
  symbol: string;
}

export interface Ssp {
  name: string;
  symbol: string;
}

export interface Mvr {
  name: string;
  symbol: string;
}

export interface Kwd {
  name: string;
  symbol: string;
}

export interface Bsd {
  name: string;
  symbol: string;
}

export interface Ngn {
  name: string;
  symbol: string;
}

export interface Gmd {
  name: string;
  symbol: string;
}

export interface Tjs {
  name: string;
  symbol: string;
}

export interface Gtq {
  name: string;
  symbol: string;
}

export interface Scr {
  name: string;
  symbol: string;
}

export interface Bmd {
  name: string;
  symbol: string;
}

export interface Clp {
  name: string;
  symbol: string;
}

export interface Bob {
  name: string;
  symbol: string;
}

export interface Jpy {
  name: string;
  symbol: string;
}

export interface Gbp {
  name: string;
  symbol: string;
}

export interface Jep {
  name: string;
  symbol: string;
}

export interface Wst {
  name: string;
  symbol: string;
}

export interface Amd {
  name: string;
  symbol: string;
}

export interface Fkp {
  name: string;
  symbol: string;
}

export interface Xof {
  name: string;
  symbol: string;
}

export interface Mdl {
  name: string;
  symbol: string;
}

export interface Kzt {
  name: string;
  symbol: string;
}

export interface Aoa {
  name: string;
  symbol: string;
}

export interface Mru {
  name: string;
  symbol: string;
}

export interface Jod {
  name: string;
  symbol: string;
}

export interface Krw {
  name: string;
  symbol: string;
}

export interface Tzs {
  name: string;
  symbol: string;
}
// ...
export interface Idd {
  root?: string;
  suffixes?: string[];
}

export interface Languages {
  eng?: string;
  spa?: string;
  mri?: string;
  nzs?: string;
  ara?: string;
  heb?: string;
  deu?: string;
  fra?: string;
  nld?: string;
  kon?: string;
  lin?: string;
  smo?: string;
  por?: string;
  mon?: string;
  bwg?: string;
  kck?: string;
  khi?: string;
  ndc?: string;
  nde?: string;
  nya?: string;
  sna?: string;
  sot?: string;
  toi?: string;
  tsn?: string;
  tso?: string;
  ven?: string;
  xho?: string;
  zib?: string;
  gil?: string;
  ton?: string;
  cha?: string;
  zdj?: string;
  msa?: string;
  pol?: string;
  jam?: string;
  dzo?: string;
  mya?: string;
  bjz?: string;
  srp?: string;
  sqi?: string;
  fas?: string;
  div?: string;
  mlt?: string;
  rus?: string;
  tgk?: string;
  crs?: string;
  aym?: string;
  grn?: string;
  que?: string;
  jpn?: string;
  nrf?: string;
  hye?: string;
  ron?: string;
  kaz?: string;
  slk?: string;
  lit?: string;
  kor?: string;
  swa?: string;
  tir?: string;
  sag?: string;
  swe?: string;
  nno?: string;
  nob?: string;
  smi?: string;
  pap?: string;
  gle?: string;
  tuk?: string;
  de?: string;
  zho?: string;
  ita?: string;
  lat?: string;
  vie?: string;
  tet?: string;
  est?: string;
  kin?: string;
  niu?: string;
  bis?: string;
  ces?: string;
  som?: string;
  lua?: string;
  dan?: string;
  fil?: string;
  pau?: string;
  lav?: string;
  bel?: string;
  tvl?: string;
  hat?: string;
  tha?: string;
  ber?: string;
  mey?: string;
  nau?: string;
  ltz?: string;
  isl?: string;
  mah?: string;
  slv?: string;
  cat?: string;
  eus?: string;
  glc?: string;
  amh?: string;
  bos?: string;
  hrv?: string;
  ell?: string;
  tur?: string;
  kir?: string;
  arc?: string;
  ckb?: string;
  khm?: string;
  kal?: string;
  bul?: string;
  afr?: string;
  her?: string;
  hgm?: string;
  kwn?: string;
  loz?: string;
  ndo?: string;
  glv?: string;
  ukr?: string;
  ssw?: string;
  cnr?: string;
  hin?: string;
  tam?: string;
  mlg?: string;
  nor?: string;
  pih?: string;
  hmo?: string;
  tpi?: string;
  urd?: string;
  kat?: string;
  nep?: string;
  fij?: string;
  hif?: string;
  gsw?: string;
  roh?: string;
  hun?: string;
  ind?: string;
  lao?: string;
  run?: string;
  nbl?: string;
  nso?: string;
  zul?: string;
  cal?: string;
  mkd?: string;
  nfr?: string;
  fao?: string;
  uzb?: string;
  sin?: string;
  ben?: string;
  prs?: string;
  pus?: string;
  rar?: string;
  fin?: string;
  tkl?: string;
  pov?: string;
  aze?: string;
  mfe?: string;
}

export interface Translations {
  ara: Ara2;
  bre: Bre;
  ces: Ces2;
  cym: Cym;
  deu: Deu2;
  est: Est2;
  fin: Fin2;
  fra: Fra2;
  hrv?: Hrv2;
  hun: Hun2;
  ita: Ita2;
  jpn?: Jpn2;
  kor: Kor2;
  nld: Nld2;
  per?: Per;
  pol: Pol2;
  por: Por2;
  rus: Rus2;
  slk: Slk2;
  spa: Spa2;
  srp: Srp2;
  swe: Swe2;
  tur: Tur2;
  urd: Urd2;
  zho?: Zho2;
}
// ...
export interface Demonyms {
  eng: Eng2;
  fra?: Fra3;
}

export interface Eng2 {
  f: string;
  m: string;
}

export interface Fra3 {
  f: string;
  m: string;
}

export interface Maps {
  googleMaps: string;
  openStreetMaps: string;
}

export interface Car {
  signs?: string[];
  side: string;
}

export interface Flags {
  png: string;
  svg: string;
  alt?: string;
}

export interface CoatOfArms {
  png?: string;
  svg?: string;
}

export interface CapitalInfo {
  latlng?: number[];
}

export interface Gini {
  "2014"?: number;
  "2016"?: number;
  "2018"?: number;
  "2011"?: number;
  "2019"?: number;
  "2006"?: number;
  "2015"?: number;
  "2004"?: number;
  "2017"?: number;
  "1999"?: number;
  "2013"?: number;
  "2009"?: number;
  "2010"?: number;
  "2008"?: number;
  "1998"?: number;
  "1992"?: number;
  "2012"?: number;
  "2003"?: number;
  "2005"?: number;
}

export interface PostalCode {
  format: string;
  regex?: string;
}